Integrate test_j2template.py into test_jinja2.py
