+++ src/uwtools/tests/config/test_jinja2.py
@@ -5,14 +5,26 @@ Tests for uwtools.utils.jinja2 module.
 
 import logging
 import os
+from types import SimpleNamespace as ns
 from unittest.mock import patch
 
 import yaml
-from pytest import fixture
+from pytest import fixture, raises
 
 from uwtools.config import jinja2
+from uwtools.config.jinja2 import J2Template
 from uwtools.logging import log
 from uwtools.tests.support import logged
+
+# Fixtures
+
+
+@fixture
+def template(tmp_path):
+    path = tmp_path / "template.jinja2"
+    with open(path, "w", encoding="utf-8") as f:
+        f.write("roses are {{roses}}, violets are {{violets}}")
+    return str(path)
 
 
 @fixture
@@ -29,12 +41,7 @@ cannot:
     return str(path)
 
 
-@fixture
-def template(tmp_path):
-    path = tmp_path / "template.jinja2"
-    with open(path, "w", encoding="utf-8") as f:
-        f.write("roses are {{roses}}, violets are {{violets}}")
-    return str(path)
+# Helpers
 
 
 def render_helper(input_file, values_file, **kwargs):
@@ -43,6 +50,16 @@ def render_helper(input_file, values_file, **kwargs):
         values_file=values_file,
         **kwargs,
     )
+
+
+def validate(template):
+    assert template._values.get("greeting") == "Hello"
+    assert template._values.get("recipient") == "the world"
+    assert template.render() == "Hello to the world"
+    assert template.undeclared_variables == {"greeting", "recipient"}
+
+
+# Tests
 
 
 def test_render(values_file, template, tmp_path):
@@ -107,3 +124,37 @@ def test__set_up_config_obj_file(values_file):
     expected = {"roses": "white", "violets": "blue", "cannot": {"override": "this"}}
     actual = jinja2._set_up_values_obj(values_file=values_file, overrides={"roses": "white"})
     assert actual == expected
+
+
+class Test_Jinja2Template:
+    """
+    Tests for class uwtools.config.jinja2.Jinja2Template.
+    """
+
+    @fixture
+    def testdata(self):
+        return ns(
+            config={"greeting": "Hello", "recipient": "the world"},
+            template="{{greeting}} to {{recipient}}",
+        )
+
+    def test_bad_args(self, testdata):
+        # It is an error to pass in neither a template path or a template string.
+        with raises(RuntimeError):
+            J2Template(testdata.config)
+
+    def test_dump(self, testdata, tmp_path):
+        path = tmp_path / "rendered.txt"
+        j2template = J2Template(testdata.config, template_str=testdata.template)
+        j2template.dump(output_path=path)
+        with open(path, "r", encoding="utf-8") as f:
+            assert f.read().strip() == "Hello to the world"
+
+    def test_render_file(self, testdata, tmp_path):
+        path = tmp_path / "template.jinja2"
+        with path.open("w", encoding="utf-8") as f:
+            print(testdata.template, file=f)
+        validate(J2Template(testdata.config, template_path=path))
+
+    def test_render_string(self, testdata):
+        validate(J2Template(testdata.config, template_str=testdata.template))
